requirements.txt
asgiref==3.5.2
blessings==1.7
bpython==0.22.1
certifi==2022.6.15
charset-normalizer==2.1.0
cloudinary==1.29.0
coreapi==2.3.3
coreschema==0.0.4
curtsies==0.3.10
cwcwidth==0.1.6
dj-database-url==0.5.0
Django==3.2.13
django-cors-headers==3.13.0
djangorestframework==3.13.1
drf-yasg==1.20.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.2.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
Pygments==2.12.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
pyxdg==0.28
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.2.0
